fix(ui): adjust GridDistortion and Hero component styling
- Fix type assertion for dataTexture in GridDistortion to prevent potential runtime errors
- Adjust grid distortion parameters for better visual effect
- Reorganize z-index layers and add pointer-events control for better interaction
- Add dark overlay directly to GridDistortion container for cleaner structure

+++ src/components/ui/GridDistortion.tsx
@@ -171,7 +171,7 @@ const GridDistortion: React.FC<GridDistortionProps> = ({
       requestAnimationFrame(animate);
       uniforms.time.value += 0.05;
 
-      const data = dataTexture.image.data;
+      const data = dataTexture.image.data as Float32Array;
       for (let i = 0; i < size * size; i++) {
         data[i * 4] *= relaxation;
         data[i * 4 + 1] *= relaxation;
